Added System() and removed RenderSystem() and UpdateSystem()
System takes a table that contains either an update() function or draw()
function. Tables that contain an update() function are added to the list
of update systems, while tables that contain a draw() function are added
to the list of draw systems.

Systems is intended to replace UpdateSystem() and RenderSystem().